requirements.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
protobuf==3.19.1
pycodestyle==2.8.0
pydantic==1.8.2
python-decouple==3.5
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.16.0
gunicorn==20.1.0
psycopg2-binary


